+++ app/src/androidTest/java/com/example/reverie2/ExampleInstrumentedTest.kt
@@ -1,8 +1,8 @@
-package com.example.reverie_v2
+package com.example.reverie2
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert.*
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
